+++ src/lib.rs
@@ -14,6 +14,7 @@ pub mod builder;
 pub mod error;
 mod event;
 mod field;
+mod internal;
 pub mod layout;
 mod render;
 pub mod style;